requirements.txt
numpy
tqdm
pygame
pillows